Move some vision-related functionality to its own module.
